add feature that enables highlight to same strs
In this commit, a somewhat shaky feature is added that will allow for
highlighting of text across all of the document. While the functionality
has been added, there is some questionable code that has been introduced
with potential bugs. Additionally, there is improvement to be made in
two cases:
1. The case where the string doesn't exactly match (i.e. "placebo " vs
"placebo").
2. Where we move/delete a span does not apply.